✓ win9x: subwnd コメント
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2426 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ WORK_01/win9x/subwnd/kdispwnd.cpp
@@ -253,7 +253,7 @@ LRESULT CKeyDisplayWnd::WindowProc(UINT nMsg, WPARAM wParam, LPARAM lParam)
 				POINT pt;
 				pt.x = GET_X_LPARAM(lParam);
 				pt.y = GET_Y_LPARAM(lParam);
-				OnRButtonDown(wParam, pt);
+				OnRButtonDown(static_cast<UINT>(wParam), pt);
 			}
 			break;
